requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
anyio==2.2.0
appdirs==1.4.3
appnope==0.1.2
argon2-cffi==21.1.0
astroid==2.5.6
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
bokeh==2.3.2
byol-pytorch==0.5.7
CacheControl==0.12.6
cachetools==4.2.2
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
cloudpickle==2.0.0
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.0
distro==1.4.0
entrypoints==0.3
flake8==4.0.1
fsspec==2021.5.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.30.0
google-auth-oauthlib==0.4.4
gradescope-utils==0.4.0
grpcio==1.37.1
gym==0.21.0
gym-super-mario-bros==7.3.2
h5py==3.3.0
html5lib==1.0.1
hydra-core==1.1.0
idna==2.8
importlib-resources==5.2.2
install==1.3.5
ipaddr==2.2.0
ipykernel==6.4.1
ipympl==0.7.0
ipyparallel==6.3.0
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.8.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyter-lsp==1.5.1
jupyter-packaging==0.9.2
jupyter-server==1.6.4
jupyter-server-mathjax==0.2.2
jupyterlab==3.2.8
jupyterlab-git==0.30.1
jupyterlab-lsp==3.10.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.0
jupyterlab-widgets==1.0.2
jupytext==1.11.2
kaggle==1.5.12
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
lightly==1.1.17
lightly-utils==0.0.2
livelossplot==0.5.4
lockfile==0.12.2
Markdown==3.3.4
markdown-it-py==1.0.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mdit-py-plugins==0.2.7
mistune==0.8.4
msgpack==0.6.2
multidict==5.1.0
nbclassic==0.2.7
nbclient==0.5.4
nbconvert==6.1.0
nbdime==3.0.0
nbformat==5.1.3
nes-py==8.1.8
nest-asyncio==1.5.1
nodeenv==1.6.0
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.0
omegaconf==2.1.0
onnx==1.9.0
onnx2pytorch==0.3.0
opencv-python==4.5.4.58
packaging==21.0
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.2
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pip==21.1.1
pkg-resources==0.0.0
plotly==4.14.3
pluggy==1.0.0
prettytable==2.1.0
progress==1.5
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.0
psutil==5.8.0
ptvsd==4.3.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pyDeprecate==0.3.0
pydocstyle==6.1.1
pyflakes==2.4.0
pyglet==1.5.11
Pygments==2.10.0
pylint==2.8.2
pynvml==8.0.4
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
python-slugify==5.0.2
pytoml==0.1.21
pytorch-lightning==1.3.7
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.1
quadprog==0.1.11
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.22.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
Send2Trash==1.8.0
setuptools==56.0.0
six==1.16.0
sklearn==0.0
smmap==4.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
tomlkit==0.7.0
torch==1.9.1+cu111
torchaudio==0.9.1
torchfile==0.1.0
torchmetrics==0.3.2
torchvision==0.10.1+cu111
tornado==6.1
tqdm==4.60.0
traitlets==5.1.0
typing-extensions==3.10.0.2
ujson==5.1.0
urllib3==1.26.5
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
wget==3.2
wheel==0.34.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xeus-python==0.12.3
yapf==0.32.0
yarl==1.6.3
zipp==3.5.0
gdown